+++ crates/toasty-driver-integration-suite/src/tests.rs
@@ -21,7 +21,6 @@ pub mod crud_sort_limit;
 pub mod crud_update_field_order;
 pub mod deferred_embed;
 pub mod deferred_field;
-pub mod dynamodb_limit_boundary;
 pub mod dynamodb_update_batch_filter;
 pub mod embed_enum_data;
 pub mod embed_enum_filter;
